+++ core/shared/src/test/scala/cats/effect/MemoizeSpec.scala
@@ -17,7 +17,7 @@
 package cats
 package effect
 
-import cats.effect.concurrent.{Deferred, Ref}
+import cats.effect.kernel.{Deferred, Ref}
 import cats.syntax.all._
 
 import org.scalacheck.Prop, Prop.forAll
