+++ core/shared/src/test/scala/cats/effect/kernel/RefSpec.scala
@@ -16,7 +16,7 @@
 
 package cats
 package effect
-package concurrent
+package kernel
 
 import cats.data.State
