+++ std/shared/src/test/scala/cats/effect/std/SyntaxSpec.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package cats.effect.concurrent
+package cats.effect.std
 
-import cats.effect.kernel.{Async, Concurrent, GenConcurrent, Sync}
+import cats.effect.kernel.{Async, Concurrent, Deferred, GenConcurrent, Ref, Sync}
 import org.specs2.mutable.Specification
 
 class SyntaxSpec extends Specification {
